backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.30.0",
    "SQLAlchemy==2.0.30",
    "psycopg2-binary==2.9.9",
    "opencv-python-headless==4.9.0.80",
    "numpy==1.26.4",
    "Pillow==10.3.0",
    "qrcode==7.4.2",
    "pyzbar==0.1.9",
    "python-multipart==0.0.9",
    "python-dotenv==1.0.1",
]
